HTTP Security added !
The API was not secured. Now, the config directory of BepInEx will have a VRisingServerApiPlugin.cfg config file with an access list consisting of comma separated values of username:password using Basic authentication.

In order to authenticate, your http request must have an authorization header with Basic {base64(username:password)} in order to get access to protected routes.

Protected routes are now defined using a boolean when declaring an http handler on a class or on a method.

Also added more info to the player details route